auth/requirements.txt
autopep8==1.6.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.1
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
toml==0.10.2
Werkzeug==2.1.2
zipp==3.8.0
